Fixed: Fix `path` command executing executable symlink target instead of symlink itself and do no execute file relative to current working directory if path does not start with `./` or `../`
For `coreutils` and `toybox`, etc, the symlink itself must be executed, like `/system/bin/ls` instead of the executable target, like `/system/bin/toybox`.

Previously, if `sudo stat` was executed and a file with the name `stat` existed in current working directory, then it would be executed instead of the one in `$PATH`, which causes nondeterministic behaviour depending on state of current directory. So we do not allow this anymore, if user wants to execute a file in current directory, it must start with `./` or `../`, like `sudo ./stat`. This behaviour is consistent with normal shell behaviour as well.