Fixed: Redirect `stderr` to `stdout`, `stdin` to `/dev/null` for `su` shells (`$SU_ENV_COMMAND`) and put all commands inside a subshell as sometimes exit code is `0` even if last command fails with a non zero exit code and command/output is out of order or stderr is not completely flushed before command exits
The reason is unknown at this time, can be tested with `sudo -vv --shell-home=$HOME/.. su`, which should always fail when `sudo_setup_sudo_shell_home_and_working_environment` returns an error and exits with exit code `1`, but doesn't.

The issue with this is that since `stderr` is redirected to `stdout`, the `su_output` that's manually logged with `echo` afterwards to `stdout` will contain both normal log entries and the errors, instead of the errors being sent to `stderr`. But this is not too big an issue as an additional error is still logged with `sudo_log_errors` to `stderr` if a command fails.